feat(core): f32 field cells and collision data — slice 2 of f32-hot-columns
WorldFields::num_values stores Option<f32> (accessors keep f64
signatures, round on store / widen on load); ColliderData centers,
points, and radii store f32 with the shared narrow-phase widening to
f64 so index and brute oracle paths agree by construction; collision
AABBs store f32 rounded OUTWARD (min toward -inf, max toward +inf) so
the broad-phase box always contains the f64 box — nearest rounding
could shrink it and prune a true hit the brute pair-set still finds.
Dyn-field and masked-update oracle compares stay bit-exact at storage
width.

Gates first-hand: 350 core tests plain + MAKU_LOWER_ORACLE=1, 6
release ignored oracle suites, scripted outcomes unchanged.

+++ crates/core/src/sim/collision.rs
@@ -866,8 +866,8 @@ fn project_colliders(
             };
             emit(ColliderData::Circle {
                 layer: *layer,
-                center: (pose.x, pose.y),
-                radius: value(*radius) * scale,
+                center: (pose.x as f32, pose.y as f32),
+                radius: (value(*radius) * scale) as f32,
             });
         }
         return;
@@ -1070,8 +1070,8 @@ impl Sim {
                 };
                 self.collider_scratch.rows.push(ColliderData::Circle {
                     layer: *layer,
-                    center: (pose.x, pose.y),
-                    radius: radius * scale,
+                    center: (pose.x as f32, pose.y as f32),
+                    radius: (radius * scale) as f32,
                 });
             }
             self.collider_scratch.finish_row(start);
